Type migrate io.sphere.sdk.queries.Query.MAX_OFFSET and MIN_OFFSTE to long.

+++ commercetools-sdk-base/src/main/java/io/sphere/sdk/queries/Query.java
@@ -4,8 +4,8 @@ import io.sphere.sdk.client.SphereRequest;
 import io.sphere.sdk.http.HttpResponse;
 
 public interface Query<T> extends SphereRequest<PagedQueryResult<T>> {
-    int MAX_OFFSET = 100000;
-    int MIN_OFFSET = 0;
+    long MAX_OFFSET = 100000;
+    long MIN_OFFSET = 0;
 
     @Override
     PagedQueryResult<T> deserialize(final HttpResponse httpResponse);
